fix: exports for docs
